+++ .gitignore
@@ -11,5 +11,6 @@ phpcs.xml
 .phpcs-cache
 /.rnd
 test/PHPMailerTest.php*nonexistent_file.txt
+test/*/PHPMailerTest.php*nonexistent_file.txt
 phpunit.xml
 /.phpunit.result.cache
